refactor(dev/cli): pruning one file
